adding theme switcher to vite app
