+++ .storybook/preview.tsx
@@ -1,5 +1,14 @@
+import React from 'react'
 import type { Preview } from "@storybook/react";
 import "../src/styles/variables.css";
+import { Decorator } from "@storybook/react";
+import { ThemeProvider, themes } from '../src/theme'
+
+const withTheme: Decorator = StoryFn => (
+  <ThemeProvider theme={themes.dark}>
+    <StoryFn />
+  </ThemeProvider>
+);
 
 const preview: Preview = {
   parameters: {
@@ -13,4 +22,5 @@ const preview: Preview = {
   },
 };
 
+export const decorators = [withTheme];
 export default preview;
