+++ src/components/index.ts
@@ -1,7 +1,7 @@
-export * from './Badge/badge'
-export * from './ButtonGroup/index'
-export * from './Card/card'
-export * from './FormField/index'
-export * from './IconButton/index'
-export * from './icons/ProfileIcon'
-export * from './icons/ProfileIcon'
+export { Badge } from './Badge/badge'
+export { ButtonGroup } from './ButtonGroup'
+export { Card } from './Card/card'
+export { TextFieldLabel } from './FormField'
+export { default as IconButton } from './IconButton'
+export { default as ProfileIcon } from './icons/ProfileIcon'
+export { default as Switch } from './Switch'
